% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_image.R
\name{measure_image}
\alias{measure_image}
\title{Measure morphometric traits of a single image}
\usage{
measure_image(
  path = system.file("inst/sample_images/example1.JPG", package = "dphrl"),
  find_eye = TRUE,
  plot_image = TRUE
)
}
\arguments{
\item{path}{path to image either absolute or relative}

\item{find_eye}{try to find eye of daphnia? boolean}

\item{plot_image}{whether or not to plot resulting image. boolean}
}
\description{
Measure morphometric traits of a single image
}
