% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_directory.R
\name{measure_directory}
\alias{measure_directory}
\title{measure all images in a directory}
\usage{
measure_directory(
  path,
  write_images = TRUE,
  eye_method = TRUE,
  scaling_factor = FALSE
)
}
\arguments{
\item{path}{path to directory. string. See Details for more info.}

\item{write_images}{write images to png? boolean}

\item{eye_method}{use eye method when possible? boolean}

\item{scaling_factor}{scale measurements to other unit. boolean. See Details for more info}
}
\description{
measure all images in a directory
}
\details{
path must contain images in either png or jpg format. note that subdirectories will be ignored (not recursive).
scaling_factor describes how many pixels occur in one other unit.
editme!
}
\examples{
\dontrun{
measure_directory("path/to/directory")
# scale images to other unit
measure_directory("path/to/directory", scaling_factor = FALSE)
}

}
